RAG_pipeline/requirements.txt

fastapi
uvicorn
torch
transformers
sentence-transformers
sentencepiece
faiss-cpu
Pillow
spacy

